+++ CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package speech
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+Unreleased
+------------------
+* Add delay functionality between end of speech and sending /robot_face?talking_finished
+
 0.1.0 (2018-07-12)
 ------------------
 * First formal release of the package